Address code review comments and refactor widget audit module
Documentation fixes:
- Update all path references from `src/ui/widget_audit.rs` to `src/ui/widget_audit/`
- Fix en-GB spelling: "Color" → "Colour", "Center" → "Centre"
- Expand UI and GPUI acronyms on first use
- Remove first-person pronoun "we" from purpose statement

BDD test improvements:
- Split oversized test module (483 lines → main.rs 166 lines, steps.rs 343 lines)
- Rename step to "all controls require action linkage" to match behaviour
- Fix `for_each_control` to fail loudly when inventory is None
- Fix `then_each_tool_has_shortcut` to validate controls exist before checking shortcuts

Code quality:
- Replace wildcard export `pub use types::*` with explicit re-exports
- Expand module doc comments for layers.rs and toolbar.rs
- Add `states` field to `ToolbarControlSpec` so each tool defines its own states
- Validate evidence file paths exist (not just prefix check) in shell seam tests

All quality gates pass:
- make check-fmt ✓
- make lint ✓
- make test ✓
