+++ src/ui/widget_audit/toolbar.rs
@@ -1,16 +1,36 @@
 //! Toolbar control definitions.
+//!
+//! This module defines the toolbar tool controls required for Phase 1 and Phase 2,
+//! including:
+//! - **Selection Tool** (`V`) — Select and move entire objects
+//! - **Direct Selection Tool** (`A`) — Select and edit anchor points or path segments
+//! - **Pen Tool** (`P`) — Create freeform paths with Bezier curves
+//! - **Rectangle Tool** (`R`) — Create rectangles and squares
+//! - **Ellipse Tool** (`O`) — Create ellipses and circles
+//! - **Line Tool** (`\`) — Create straight line segments
+//! - **Type Tool** (`T`) — Create and edit text objects (Phase 2)
+//!
+//! These controls are used by the widget capability audit to verify that
+//! the Phase 0 shell implements the required toolbar functionality.
+//! Each tool specifies its own states via `ControlStates`, allowing tools
+//! like the Type Tool to define custom state lists as needed.
 
 use super::types::{
     AccessibilityRequirements, ActionCommandLinkage, ControlStates, ControlSurface,
     CurrentShellEvidence, KeyboardRequirements, Phase, RequiredControl, RequirementSource, UserJob,
 };
 
+/// Specification for a toolbar tool control.
+///
+/// Each tool defines its own states via the `states` field, allowing
+/// customization for tools with unique state requirements.
 #[derive(Clone, Copy)]
 struct ToolbarControlSpec {
     name: &'static str,
     phase: Phase,
     description: &'static str,
     shortcut: &'static str,
+    states: &'static [&'static str],
     a11y_notes: &'static str,
     action_name: &'static str,
     roadmap_ref: &'static str,
@@ -29,7 +49,7 @@ fn make_toolbar_tool(spec: ToolbarControlSpec) -> RequiredControl {
             description: spec.description,
         },
         states: ControlStates {
-            states: vec!["selected", "disabled", "focused"],
+            states: spec.states.to_vec(),
         },
         keyboard: KeyboardRequirements {
             shortcut: Some(spec.shortcut),
@@ -59,6 +79,9 @@ fn make_toolbar_tool(spec: ToolbarControlSpec) -> RequiredControl {
     }
 }
 
+/// Standard tool states used by most toolbar tools.
+const STANDARD_TOOL_STATES: &[&str] = &["selected", "disabled", "focused"];
+
 /// Toolbar controls (7 tools).
 pub(super) fn controls() -> Vec<RequiredControl> {
     TOOL_SPECS.iter().copied().map(make_toolbar_tool).collect()
@@ -70,6 +93,7 @@ const TOOL_SPECS: &[ToolbarControlSpec] = &[
         phase: Phase::Phase1,
         description: "Select and move entire objects",
         shortcut: "V",
+        states: STANDARD_TOOL_STATES,
         a11y_notes: "Must announce tool activation",
         action_name: "ActivateSelectionTool",
         roadmap_ref: "1.2.2",
@@ -83,6 +107,7 @@ const TOOL_SPECS: &[ToolbarControlSpec] = &[
         phase: Phase::Phase1,
         description: "Select and edit individual anchor points or path segments",
         shortcut: "A",
+        states: STANDARD_TOOL_STATES,
         a11y_notes: "Must announce tool activation and distinguish from Selection Tool",
         action_name: "ActivateDirectSelectionTool",
         roadmap_ref: "1.2.2",
@@ -96,6 +121,7 @@ const TOOL_SPECS: &[ToolbarControlSpec] = &[
         phase: Phase::Phase1,
         description: "Create freeform paths with Bezier curves",
         shortcut: "P",
+        states: STANDARD_TOOL_STATES,
         a11y_notes: "Must announce tool activation",
         action_name: "ActivatePenTool",
         roadmap_ref: "1.1.4",
@@ -109,6 +135,7 @@ const TOOL_SPECS: &[ToolbarControlSpec] = &[
         phase: Phase::Phase1,
         description: "Create rectangles and squares",
         shortcut: "R",
+        states: STANDARD_TOOL_STATES,
         a11y_notes: "Must announce tool activation",
         action_name: "ActivateRectangleTool",
         roadmap_ref: "1.1.1",
@@ -122,6 +149,7 @@ const TOOL_SPECS: &[ToolbarControlSpec] = &[
         phase: Phase::Phase1,
         description: "Create ellipses and circles",
         shortcut: "O",
+        states: STANDARD_TOOL_STATES,
         a11y_notes: "Must announce tool activation",
         action_name: "ActivateEllipseTool",
         roadmap_ref: "1.1.2",
@@ -135,6 +163,7 @@ const TOOL_SPECS: &[ToolbarControlSpec] = &[
         phase: Phase::Phase1,
         description: "Create straight line segments",
         shortcut: "\\",
+        states: STANDARD_TOOL_STATES,
         a11y_notes: "Must announce tool activation",
         action_name: "ActivateLineTool",
         roadmap_ref: "1.1.3",
@@ -148,6 +177,7 @@ const TOOL_SPECS: &[ToolbarControlSpec] = &[
         phase: Phase::Phase2,
         description: "Create and edit text objects",
         shortcut: "T",
+        states: STANDARD_TOOL_STATES,
         a11y_notes: "Must announce tool activation and text editing mode",
         action_name: "ActivateTypeTool",
         roadmap_ref: "2.1.1",
